requirements.txt
gradio
openai
Pillow
qrcode